lyceum/requirements.txt
asgiref==3.5.0
black @ git+https://github.com/psf/black@14d84ba2e96c5ca1351b8fe4d0d415cc148f4117
click==8.0.4
colorama==0.4.4
Django==3.2.12
django-cleanup==6.0.0
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-extensions==3.1.5
django-widget-tweaks==1.4.12
flake8==4.0.1
flake8-import-order==0.18.1
isort==5.10.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.1
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2021.3
sorl-thumbnail==12.8.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
